fix: refactor some of the buttons
